+++ platform/src/test/java/org/bytedeco/javacv/FrameGrabberChangingResolutionTest.java
@@ -53,7 +53,7 @@ public class FrameGrabberChangingResolutionTest {
         recorder.setVideoCodecName("libx264rgb");
         recorder.setVideoQuality(0); // lossless
         recorder.setFrameRate(30);
-        recorder.start();
+        recorder.startUnsafe();
 
         Frame[] frames = new Frame[60];
         for (int n = 0; n < frames.length; n++) {
@@ -74,10 +74,10 @@ public class FrameGrabberChangingResolutionTest {
     }
 
     final public void setupUDPSender(final int x, final int y, final int bandwidth, final int count) throws IOException {
-        final FrameGrabber fg = new FFmpegFrameGrabber(tempFile);
+        final FFmpegFrameGrabber fg = new FFmpegFrameGrabber(tempFile);
         fg.setFrameRate(30);
 
-        final FrameRecorder fr = new FFmpegFrameRecorder("udp://127.0.0.1:2345", 0);
+        final FFmpegFrameRecorder fr = new FFmpegFrameRecorder("udp://127.0.0.1:2345", 0);
         fr.setVideoCodecName("mpeg2video");
         fr.setFormat("mpegts");
 
@@ -87,8 +87,8 @@ public class FrameGrabberChangingResolutionTest {
 
         fr.setFrameRate(30);
 
-        fg.start();
-        fr.start();
+        fg.startUnsafe();
+        fr.startUnsafe();
 
         final boolean[] b = new boolean[1];
         Thread t = new Thread() {
@@ -132,10 +132,10 @@ public class FrameGrabberChangingResolutionTest {
         Thread t = new Thread() {
 
             public void run() {
-                FrameGrabber fg = new FFmpegFrameGrabber("udp://127.0.0.1:2345");
+                FFmpegFrameGrabber fg = new FFmpegFrameGrabber("udp://127.0.0.1:2345");
                 fg.setFrameRate(30);
 
-                FrameRecorder fr = new FFmpegFrameRecorder(tempTargetFile, 0);
+                FFmpegFrameRecorder fr = new FFmpegFrameRecorder(tempTargetFile, 0);
                 fr.setVideoCodecName("mpeg2video");
                 fr.setFormat("mpegts");
 
@@ -146,8 +146,8 @@ public class FrameGrabberChangingResolutionTest {
                 fr.setFrameRate(30);
 
                 try {
-                    fg.start();
-                    fr.start();
+                    fg.startUnsafe();
+                    fr.startUnsafe();
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
@@ -240,7 +240,7 @@ public class FrameGrabberChangingResolutionTest {
         try {
             FFmpegFrameGrabber grabber = new FFmpegFrameGrabber(new FileInputStream(tempTargetFile));
             grabber.setSampleMode(FrameGrabber.SampleMode.FLOAT);
-            grabber.start();
+            grabber.startUnsafe();
 
             int n = 0;
             Frame frame2;
@@ -257,7 +257,6 @@ public class FrameGrabberChangingResolutionTest {
             assertTrue(n > 300);
             assertTrue(n <= 480);
             assertEquals(null, grabber.grab());
-            grabber.restart();
             grabber.stop();
             grabber.release();
         } catch (Exception e) {