+++ src/main/java/org/bytedeco/javacv/FFmpegFrameRecorder.java
@@ -171,21 +171,24 @@ public class FFmpegFrameRecorder extends FrameRecorder {
     public FFmpegFrameRecorder(OutputStream outputStream, int audioChannels) {
         this(outputStream.toString(), audioChannels);
         this.outputStream = outputStream;
+        this.closeOutputStream = true;
     }
     public FFmpegFrameRecorder(OutputStream outputStream, int imageWidth, int imageHeight) {
         this(outputStream.toString(), imageWidth, imageHeight);
         this.outputStream = outputStream;
+        this.closeOutputStream = true;
     }
     public FFmpegFrameRecorder(OutputStream outputStream, int imageWidth, int imageHeight, int audioChannels) {
         this(outputStream.toString(), imageWidth, imageHeight, audioChannels);
         this.outputStream = outputStream;
+        this.closeOutputStream = true;
     }
     public void release() throws Exception {
         // synchronized (org.bytedeco.javacpp.avcodec.class) {
         releaseUnsafe();
         // }
     }
-    void releaseUnsafe() throws Exception {
+    public void releaseUnsafe() throws Exception {
         /* close each codec */
         if (video_c != null) {
             avcodec_free_context(video_c);
@@ -274,7 +277,9 @@ public class FFmpegFrameRecorder extends FrameRecorder {
 
         if (outputStream != null) {
             try {
-                outputStream.close();
+                if (closeOutputStream) {
+                    outputStream.close();
+                }
             } catch (IOException ex) {
                 throw new Exception("Error on OutputStream.close(): ", ex);
             } finally {
@@ -323,6 +328,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
     }
 
     private OutputStream outputStream;
+    private boolean closeOutputStream;
     private AVIOContext avio;
     private String filename;
     private AVFrame picture, tmp_picture;
@@ -348,6 +354,13 @@ public class FFmpegFrameRecorder extends FrameRecorder {
     private AVPacket video_pkt, audio_pkt;
     private int[] got_video_packet, got_audio_packet;
     private AVFormatContext ifmt_ctx;
+
+    public boolean isCloseOutputStream() {
+        return closeOutputStream;
+    }
+    public void setCloseOutputStream(boolean closeOutputStream) {
+        this.closeOutputStream = closeOutputStream;
+    }
 
     @Override public int getFrameNumber() {
         return picture == null ? super.getFrameNumber() : (int)picture.pts();
@@ -375,7 +388,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
         // }
     }
 
-    void startUnsafe() throws Exception {
+    public void startUnsafe() throws Exception {
         int ret;
         picture = null;
         tmp_picture = null;
@@ -471,18 +484,18 @@ public class FFmpegFrameRecorder extends FrameRecorder {
 
             /* add a video output stream */
             if ((video_st = avformat_new_stream(oc, null)) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avformat_new_stream() error: Could not allocate video stream.");
             }
 
             if ((video_c = avcodec_alloc_context3(video_codec)) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avcodec_alloc_context3() error: Could not allocate video encoding context.");
             }
 
             if (inpVideoStream != null) {
                 if ((ret = avcodec_copy_context(video_st.codec(), inpVideoStream.codec())) < 0) {
-                    release();
+                    releaseUnsafe();
                     throw new Exception("avcodec_copy_context() error:\tFailed to copy context from input to output stream codec context");
                 }
 
@@ -609,12 +622,12 @@ public class FFmpegFrameRecorder extends FrameRecorder {
             oformat.audio_codec(audio_codec.id());
 
             if ((audio_st = avformat_new_stream(oc, null)) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avformat_new_stream() error: Could not allocate audio stream.");
             }
 
             if ((audio_c = avcodec_alloc_context3(audio_codec)) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avcodec_alloc_context3() error: Could not allocate audio encoding context.");
             }
 
@@ -699,7 +712,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
             }
             /* open the codec */
             if ((ret = avcodec_open2(video_c, video_codec, options)) < 0) {
-                release();
+                releaseUnsafe();
                 av_dict_free(options);
                 throw new Exception("avcodec_open2() error " + ret + ": Could not open video codec.");
             }
@@ -719,27 +732,27 @@ public class FFmpegFrameRecorder extends FrameRecorder {
 
             /* allocate the encoded raw picture */
             if ((picture = av_frame_alloc()) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("av_frame_alloc() error: Could not allocate picture.");
             }
             picture.pts(0); // magic required by libx264
 
             int size = av_image_get_buffer_size(video_c.pix_fmt(), video_c.width(), video_c.height(), 1);
             if ((picture_buf = new BytePointer(av_malloc(size))).isNull()) {
-                release();
+                releaseUnsafe();
                 throw new Exception("av_malloc() error: Could not allocate picture buffer.");
             }
 
             /* if the output format is not equal to the image format, then a temporary
                picture is needed too. It is then converted to the required output format */
             if ((tmp_picture = av_frame_alloc()) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("av_frame_alloc() error: Could not allocate temporary picture.");
             }
 
             /* copy the stream parameters to the muxer */
             if ((ret = avcodec_parameters_from_context(video_st.codecpar(), video_c)) < 0) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avcodec_parameters_from_context() error: Could not copy the video stream parameters.");
             }
 
@@ -760,7 +773,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
             }
             /* open the codec */
             if ((ret = avcodec_open2(audio_c, audio_codec, options)) < 0) {
-                release();
+                releaseUnsafe();
                 av_dict_free(options);
                 throw new Exception("avcodec_open2() error " + ret + ": Could not open audio codec.");
             }
@@ -801,14 +814,14 @@ public class FFmpegFrameRecorder extends FrameRecorder {
 
             /* allocate the audio frame */
             if ((frame = av_frame_alloc()) == null) {
-                release();
+                releaseUnsafe();
                 throw new Exception("av_frame_alloc() error: Could not allocate audio frame.");
             }
             frame.pts(0); // magic required by libvorbis and webm
 
             /* copy the stream parameters to the muxer */
             if ((ret = avcodec_parameters_from_context(audio_st.codecpar(), audio_c)) < 0) {
-                release();
+                releaseUnsafe();
                 throw new Exception("avcodec_parameters_from_context() error: Could not copy the audio stream parameters.");
             }
 
@@ -828,7 +841,7 @@ public class FFmpegFrameRecorder extends FrameRecorder {
         if (outputStream == null && (oformat.flags() & AVFMT_NOFILE) == 0) {
             AVIOContext pb = new AVIOContext(null);
             if ((ret = avio_open2(pb, filename, AVIO_FLAG_WRITE, null, options)) < 0) {
-                release();
+                releaseUnsafe();
                 av_dict_free(options);
                 throw new Exception("avio_open2 error() error " + ret + ": Could not open '" + filename + "'");
             }
@@ -845,6 +858,23 @@ public class FFmpegFrameRecorder extends FrameRecorder {
 
         if (av_log_get_level() >= AV_LOG_INFO) {
             av_dump_format(oc, 0, filename, 1);
+        }
+    }
+
+    public void flush() throws Exception {
+        synchronized (oc) {
+            /* flush all the buffers */
+            while (video_st != null && ifmt_ctx == null && recordImage(0, 0, 0, 0, 0, AV_PIX_FMT_NONE, (Buffer[])null));
+            while (audio_st != null && ifmt_ctx == null && recordSamples(0, 0, (Buffer[])null));
+
+            if (interleaved && video_st != null && audio_st != null) {
+                av_interleaved_write_frame(oc, null);
+            } else {
+                av_write_frame(oc, null);
+            }
+
+            /* write the trailer, if any */
+            av_write_trailer(oc);
         }
     }