+++ src/main/java/org/bytedeco/javacv/OpenCVFrameRecorder.java
@@ -98,6 +98,9 @@ public class OpenCVFrameRecorder extends FrameRecorder {
         return videoCodec;
     }
 
+    public void flush() throws Exception {
+    }
+
     public void stop() throws Exception {
         release();
     }