+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 
+ * Use ModiTect to compile `module-info.java` with JDK 8 and preserve backward compatibility
+ * Add `FFmpegFrameRecorder.closeOutputStream` and `FFmpegFrameGrabber.closeInputStream` properties to leave streams opened ([issue #1149](https://github.com/bytedeco/javacv/issues/1149))
+ * Add `FFmpegFrameRecorder.flush()` method that does not release the stream ([issue #1149](https://github.com/bytedeco/javacv/issues/1149))
+ * Readd `synchronized` blocks for `FFmpegFrameGrabber` and `FFmpegFrameRecorder`, but make unsafe methods public ([issue #1139](https://github.com/bytedeco/javacv/issues/1139))
  * Allocate native memory for `Frame` using `Pointer` to allow deallocation with `PointerScope` ([issue #1152](https://github.com/bytedeco/javacv/issues/1152))
  * Add `module-info.java` and depend on modularized JavaCPP Presets to comply with JPMS
- * Upgrade dependencies for FFmpeg 4.1.1 and libfreenect 0.5.7
+ * Upgrade dependencies for FFmpeg 4.1.2 and libfreenect 0.5.7
  * Allow allocation of `Frame` images with custom strides
  * Take into account `Bitmap.getRowBytes()` in `AndroidFrameConverter.convert(Bitmap)` ([issue #1143](https://github.com/bytedeco/javacv/issues/1143))
  * Add `static { Loader.load(); }` in `LeptonicaFrameConverter` and `OpenCVFrameConverter` to prevent link errors ([issue #1128](https://github.com/bytedeco/javacv/issues/1128))